+++ packages/@expo/cli/e2e/fixtures/with-hmr-env-vars/App.js
@@ -1,0 +1,8 @@
+import '@expo/metro-runtime';
+
+import React from 'react';
+import { Text } from 'react-native';
+
+export default function App() {
+  return <Text testID="env-var">{process.env.EXPO_PUBLIC_VALUE}</Text>;
+}