+++ packages/babel-preset-expo/src/__tests__/client-module-proxy-plugin.test.ts
@@ -342,7 +342,7 @@ export default Svg;
     `;
 
     const contents = babel.transform(sourceCode, options);
-    expect(contents?.metadata).toEqual({ hasCjsExports: false });
+    expect(contents?.metadata).toEqual({ hasCjsExports: false, publicEnvVars: [] });
 
     expect(contents?.code).not.toMatch('react-server-dom-webpack');
   });